Messages Controller, API Response, Error Logging

+++ WKLNAMA/Controllers/MessageController.cs
@@ -1,10 +1,12 @@
-﻿using Business.Services;
+﻿using Business.Helpers.Attributes;
+using Business.Services;
 using Business.ViewModels;
 using Data.DomainModels;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.SignalR;
+using Newtonsoft.Json.Linq;
 using System.CodeDom;
 using System.Net;
 using WKLNAMA.AppHub;
@@ -14,7 +16,7 @@ namespace WKLNAMA.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    public class MessageController : BaseController<Message>
+    public class MessageController : BaseController<Data.DomainModels.Message>
     {
         private readonly IMessageRepository _message;
         private readonly IHttpContextAccessor httpContextAccessor;
@@ -30,45 +32,90 @@ namespace WKLNAMA.Controllers
         [HttpPost("GetPrivateChat")] 
         public async Task<ActionResult> GetPrivateChat(MessageVm userDetails) 
         {
-            if (userDetails.FromUserId == UserModel.UserId) 
+
+            return await APIResponse(async () =>
             {
-                var messages = await _message.GetPrivateChat(userDetails);
-                return Ok(messages);
-            }
-            return BadRequest();
+                 apiResponse.Data = null;
+                apiResponse.Message = HttpStatusCode.NotFound.ToString();
+                apiResponse.HttpStatusCode = HttpStatusCode.NotFound;
+
+                if (userDetails.FromUserId == UserModel.UserId)
+                {
+                    var messages = await _message.GetPrivateChat(userDetails);
+                    apiResponse.Message = HttpStatusCode.OK.ToString();
+                    apiResponse.HttpStatusCode = HttpStatusCode.OK;
+                    apiResponse.Success = true;
+                    apiResponse.Data = messages;
+                }
+
+                return Ok(apiResponse);
+
+            });
         }
 
         [HttpPost("MarkRead")]
         public async Task<ActionResult> MarkMessagesRead(List<MessageVm> messages) 
         {
-            if (messages?.FirstOrDefault()?.ToUserId == UserModel.UserId)
+            return await APIResponse(async () =>
             {
-                 await _message.MarkRead(messages);
-                return Ok("Messages has been marked as read");
-            }
-            return BadRequest();
+                apiResponse.Data = null;
+                apiResponse.Message = HttpStatusCode.NotFound.ToString();
+                apiResponse.HttpStatusCode = HttpStatusCode.NotFound;
+
+                if (messages?.FirstOrDefault()?.ToUserId == UserModel.UserId)
+                {
+                    await _message.MarkRead(messages);
+
+                    apiResponse.Message = HttpStatusCode.OK.ToString();
+                    apiResponse.HttpStatusCode = HttpStatusCode.OK;
+                    apiResponse.Success = true;
+                    apiResponse.Data = "Messages has been marked as read";
+                }
+                return Ok(apiResponse);
+            });
         }
 
 
         [HttpPost("GetAllUnReadMessageCount")] 
         public async Task<ActionResult> GetAllUnReadMessageCount(long userId)
         {
-            if (userId == UserModel.UserId)
-            {
-                await _message.GetAllUnReadMessagesCount(userId);
-                return Ok("Messages has been marked as read");
-            }
-            return BadRequest();
+          return  await APIResponse(async () => {
+
+              apiResponse.HttpStatusCode = HttpStatusCode.BadRequest;
+              apiResponse.Message = HttpStatusCode.BadRequest.ToString();
+
+              if (userId == UserModel.UserId)
+              {
+                var messageCount= await _message.GetAllUnReadMessagesCount(userId);
+                  apiResponse.Message = HttpStatusCode.OK.ToString();
+                  apiResponse.HttpStatusCode = HttpStatusCode.OK;
+                  apiResponse.Success = true;
+                  apiResponse.Data = messageCount;
+              }
+               return Ok(apiResponse);
+          });
+           
         }
         [HttpPost("GetUnReadMessageCountByUser")]
         public async Task<ActionResult> GetUnReadMessageCountByUser(MessageVm messages)
         {
-            if (messages?.ToUserId == UserModel.UserId)
+          return  await APIResponse(async () => 
             {
-                await _message.GetUnReadMessagesCountByUser(messages);
-                return Ok("Messages has been marked as read");
-            }
-            return BadRequest();
+                apiResponse.HttpStatusCode = HttpStatusCode.BadRequest;
+                apiResponse.Message = HttpStatusCode.BadRequest.ToString();
+
+                if (messages?.ToUserId == UserModel.UserId)
+                {
+                  var userUnreadMessages=   await _message.GetUnReadMessagesCountByUser(messages);
+
+                    apiResponse.Message = HttpStatusCode.OK.ToString();
+                    apiResponse.HttpStatusCode = HttpStatusCode.OK;
+                    apiResponse.Success = true;
+                    apiResponse.Data = userUnreadMessages;
+                }
+                return Ok(apiResponse);
+            });
+
         }
 
 
@@ -79,35 +126,75 @@ namespace WKLNAMA.Controllers
         [HttpPost("Send")]
         public async  Task<ActionResult> Send(MessageVm message)
         {
+            return await APIResponse(async () =>
+            {
+                apiResponse.HttpStatusCode = HttpStatusCode.BadRequest;
+                apiResponse.Message = HttpStatusCode.BadRequest.ToString();
 
-           await _message.Create(message);
+                if (message?.ToUserId == UserModel.UserId)
+                {
+                    var messages = await _message.Create(message);
 
-            await _chatHub.Clients.User(message.ToUserId.ToString()).DirectMessage(message);
+                    await _chatHub.Clients.User(message.ToUserId.ToString()).DirectMessage(message);
 
-            return NoContent();
-            
+                    apiResponse.Message = HttpStatusCode.OK.ToString();
+                    apiResponse.HttpStatusCode = HttpStatusCode.OK;
+                    apiResponse.Success = true;
+                    apiResponse.Data = messages;
+                }
+                return Ok(apiResponse);
+            });
         }
         [HttpPost("Delete")]
         public async Task<ActionResult> Delete(MessageVm message)
         {
-            await _message.Delete(message.Id);
-            await _message.SaveAsync();
+            return await APIResponse(async () => {
 
-            await _chatHub.Clients.User(message.FromUserId.ToString()).DeleteMessage( "Message has bee deleted", message.Id.ToString());
+                apiResponse.Message = HttpStatusCode.BadRequest.ToString();
+                apiResponse.HttpStatusCode = HttpStatusCode.BadRequest;
 
-            return NoContent();
+                if (message?.ToUserId == UserModel.UserId)
+                {
+                    await _message.Delete(message.Id);
+                    await _message.SaveAsync();
+
+                    apiResponse.Message = HttpStatusCode.OK.ToString();
+                    apiResponse.HttpStatusCode = HttpStatusCode.OK;
+                    apiResponse.Success = true;
+                    apiResponse.Data = "Message has been deleted";
+
+                    await _chatHub.Clients.User(message.FromUserId.ToString()).DeleteMessage("Message has bee deleted", message.Id.ToString());
+                }
+                return Ok(apiResponse);
+
+            });
+
+          
         }
 
         [HttpPost("Reply")]
         public async Task<ActionResult> Reply(MessageVm message)
         {
-             
-            await _message.Create(message);
-            await _message.SaveAsync();
+            return await APIResponse(async () => {
 
-            await _chatHub.Clients.User(message.ToUserId.ToString()).DirectMessage(message);
 
-            return NoContent();
+                apiResponse.Message = HttpStatusCode.BadRequest.ToString();
+                apiResponse.HttpStatusCode = HttpStatusCode.BadRequest;
+
+                if (message?.ToUserId == UserModel.UserId)
+                {
+                  var _messageSent= await _message.Create(message);
+                    await _message.SaveAsync();
+
+                    await _chatHub.Clients.User(message.ToUserId.ToString()).DirectMessage(_messageSent);
+
+                    apiResponse.Message = HttpStatusCode.OK.ToString(); 
+                    apiResponse.HttpStatusCode = HttpStatusCode.OK;
+                    apiResponse.Success = true;
+                    apiResponse.Data = _messageSent;
+                }
+                return Ok(apiResponse);
+            });
         }
 
         #region Hide Override Methods
@@ -116,14 +203,14 @@ namespace WKLNAMA.Controllers
         public override async Task<ActionResult> GetAll() => await Task.FromResult(NoContent());
 
         [NonAction]
-        public override async Task<ActionResult> Post(Message T) => await Task.FromResult(NoContent());
+        public override async Task<ActionResult> Post(Data.DomainModels.Message T) => await Task.FromResult(NoContent());
         [NonAction]
         public async override Task<ActionResult> GetById(Object _viewModel) =>  await Task.FromResult(NoContent());
 
         [NonAction]
         public async override Task<ActionResult> Delete(object id) => await Task.FromResult( NoContent());
         [NonAction]
-        public async override Task<ActionResult> Update(Message _object)=> await Task.FromResult(NoContent());
+        public async override Task<ActionResult> Update(Data.DomainModels.Message _object)=> await Task.FromResult(NoContent());
         #endregion
 
     }
